requirements.txt
apache-airflow==2.8.1
mlflow==2.10.2
pandas==2.2.0
scikit-learn==1.4.0
joblib==1.3.2
pyarrow==15.0.0